requirements.txt
channels==4.0.0
channels-redis==4.2.0
coverage==7.3.4
daphne==4.0.0
Django==4.2.6
django-extensions==3.2.3
djangorestframework==3.14.0
pillow==10.2.0
redis==5.0.1